Renaming Branch PR info to Trigger Info, added Tag to it as well
After some thinking this column is meant to provide info about the Trigger type & related info,
e.g. in case of a push the branch, if tag trigger happens then the tag,
and in case of a PR build the PR ID, source & target branches.
So, renamed the related functions to Trigger Info, as well as added Tag
as presented info